more on handling errors and customizing error msgs

+++ controller/stables.js
@@ -1,26 +1,29 @@
-const stablesSchema = require("../Schema/stables");
-const stablesOrderSchema = require("../Schema/order");
+const stablesSchema = require("../Schema/stables")
+const stablesOrderSchema = require("../Schema/order")
+const { IDNotFound, DuplicateError } = require('../errors/customErrors')
 
 // GET/stables/inventory
 const getInventory = async (req, res) => {
-  const stablesInventory = await stablesSchema.find();
+  const stablesInventory = await stablesSchema.find()
   const numberOfHorses = stablesInventory.reduce((total, stablesSchema) => total + stablesSchema.numberOfHorses, 0)
 
-  res.status(200).json(numberOfHorses);
-};
+  res.status(200).json(numberOfHorses)
+}
 
 // GET/allstables/
 const getAll = async (req, res) => {
-  const allStables = await stablesSchema.find();
-  res.status(200).json(allStables);
-};
+  const allStables = await stablesSchema.find()
+  res.status(200).json(allStables)
+}
 
 // GET/stables/{stablesId}
 const getOne = async (req, res) => {
-  const stable = await stablesSchema.findById(req.params.id)
+  const { id } = req.params
+  const stable = await stablesSchema.findById(id)
 
   if (!stable) {
-    throw new Error('ID not found')
+    const error = new IDNotFound(`Stable with ID ${id} not found`)
+    throw error
   }
 
   res.status(200).json(stable)
@@ -37,7 +40,8 @@ const orderOne = async (req, res) => {
   const existingOrder = await stablesOrderSchema.findOne({ horseId: orderReq.horseId })
 
   if (existingOrder) {
-    throw new Error(`Horse with ID ${orderReq.horseId} already exists`)
+    const error = new DuplicateError(`Horse with ID ${orderReq.horseId} already exists`)
+    throw error
   }
   const newOrder = await orderReq.save()
   res.status(201).json(newOrder._id)
@@ -64,7 +68,8 @@ const createOne = async (req, res) => {
   })
 
   if (createdStable) {
-    throw new Error(`Stable ${stable.name} already on DB`)
+    const error = new DuplicateError(`Stable ${stable.name} already on DB`)
+    throw error
   }
   const newStable = await stable.save()
   res.status(201).json(newStable._id)
@@ -79,7 +84,8 @@ const updateOne = async (req, res) => {
   // Check if stable with given ID exists
   const existingStableById = await stablesSchema.findById(id)
   if (!existingStableById) {
-    throw new Error(`Stable with ID ${id} not found`)
+    const error = new IDNotFound(`Stable with ID ${id} not found`)
+    throw error
   }
 
   // Check if stable with same name, location, and owner already exists
@@ -93,11 +99,14 @@ const updateOne = async (req, res) => {
   })
 
   if (existingStable && existingStable.id.toString() !== id) {
-    throw new Error(JSON.stringify({ error: `This update creates a duplicate of another saved stable data` }))
+    const error = new DuplicateError(`This update creates a duplicate of another saved stable data`)
+    throw error
   }
 
   if (existingStable && existingStable._id.toString() == id) {
-    throw new Error(`The data you entered has the same data to the original one`)
+    const error = new DuplicateError(`The data you entered has the same data to the original one`)
+    throw error
+
   }
 
   // Update stable
@@ -110,19 +119,22 @@ const updateOne = async (req, res) => {
   }, { new: true, runValidators: true }
   )
 
-  if (!updatedStable) {
-    throw new Error(`Stable with ID ${id} not found`)
-  }
-
-  res.status(200).json(updatedStable);
+  res.status(200).json(updatedStable)
 }
 
 // DELETE/stables/{stablesId}
 const deleteOne = async (req, res) => {
+  const { id } = req.params
+  const stable = await stablesSchema.findById(id)
+
+  if (!stable) {
+    const error = new IDNotFound(`Stable with ID ${id} not found`)
+    throw error
+  }
 
   const removedStables = await stablesSchema.deleteOne({ _id: req.params.id })
   res.status(200).json(removedStables)
 
 }
 
-module.exports = { getInventory, getAll, getOne, createOne, orderOne, updateOne, deleteOne };
+module.exports = { getInventory, getAll, getOne, createOne, orderOne, updateOne, deleteOne }
